refactor(trie): extract cursor database implementation to a new `trie-db` crate
- replace `TrieWalker` dependency on database with a trait
- use custom associated error type instead of `DatabaseError` everywhere
- remove dependencies on `reth-db` and `reth-db-api` from `reth-trie`
