+++ bin/reth/Cargo.toml
@@ -54,6 +54,7 @@ reth-basic-payload-builder.workspace = true
 reth-static-file.workspace = true
 reth-static-file-types = { workspace = true, features = ["clap"] }
 reth-trie = { workspace = true, features = ["metrics"] }
+reth-trie-db = { workspace = true, features = ["metrics"] }
 reth-node-api.workspace = true
 reth-node-optimism = { workspace = true, optional = true, features = [
     "optimism",
