+++ bin/reth/src/commands/debug_cmd/build_block.rs
@@ -37,6 +37,7 @@ use reth_transaction_pool::{
     blobstore::InMemoryBlobStore, BlobStore, EthPooledTransaction, PoolConfig, TransactionOrigin,
     TransactionPool, TransactionValidationTaskExecutor,
 };
+use reth_trie_db::state::state_root_with_updates;
 use std::{path::PathBuf, str::FromStr, sync::Arc};
 use tracing::*;
 
@@ -283,9 +284,10 @@ impl Command {
                 debug!(target: "reth::cli", ?execution_outcome, "Executed block");
 
                 let hashed_post_state = execution_outcome.hash_state_slow();
-                let (state_root, trie_updates) = execution_outcome
-                    .hash_state_slow()
-                    .state_root_with_updates(provider_factory.provider()?.tx_ref())?;
+                let (state_root, trie_updates) = state_root_with_updates(
+                    execution_outcome.hash_state_slow(),
+                    provider_factory.provider()?.tx_ref(),
+                )?;
 
                 if state_root != block_with_senders.state_root {
                     eyre::bail!(
