+++ bin/reth/src/commands/debug_cmd/in_memory_merkle.rs
@@ -27,6 +27,7 @@ use reth_revm::database::StateProviderDatabase;
 use reth_stages::StageId;
 use reth_tasks::TaskExecutor;
 use reth_trie::StateRoot;
+use reth_trie_db::{state::state_root_with_updates, trie::StateRootDb};
 use std::{path::PathBuf, sync::Arc};
 use tracing::*;
 
@@ -149,7 +150,7 @@ impl Command {
 
         // Unpacked `BundleState::state_root_slow` function
         let (in_memory_state_root, in_memory_updates) =
-            execution_outcome.hash_state_slow().state_root_with_updates(provider.tx_ref())?;
+            state_root_with_updates(execution_outcome.hash_state_slow(), provider.tx_ref())?;
 
         if in_memory_state_root == block.state_root {
             info!(target: "reth::cli", state_root = ?in_memory_state_root, "Computed in-memory state root matches");
