+++ crates/stages/stages/src/stages/merkle.rs
@@ -15,6 +15,7 @@ use reth_stages_api::{
     StageCheckpoint, StageError, StageId, UnwindInput, UnwindOutput,
 };
 use reth_trie::{IntermediateStateRootState, StateRoot, StateRootProgress, StoredSubNode};
+use reth_trie_db::{trie::StateRootDb, TxRefWrapper};
 use std::fmt::Debug;
 use tracing::*;
 
@@ -217,7 +218,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
                 })?;
             match progress {
                 StateRootProgress::Progress(state, hashed_entries_walked, updates) => {
-                    updates.write_to_database(tx)?;
+                    updates.write_to_database(&TxRefWrapper::from(tx))?;
 
                     let checkpoint = MerkleCheckpoint::new(
                         to_block,
@@ -237,7 +238,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
                     })
                 }
                 StateRootProgress::Complete(root, hashed_entries_walked, updates) => {
-                    updates.write_to_database(tx)?;
+                    updates.write_to_database(&TxRefWrapper::from(tx))?;
 
                     entities_checkpoint.processed += hashed_entries_walked as u64;
 
@@ -252,7 +253,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
                         error!(target: "sync::stages::merkle", %e, ?current_block_number, ?to_block, "Incremental state root failed! {INVALID_STATE_ROOT_ERROR_MESSAGE}");
                         StageError::Fatal(Box::new(e))
                     })?;
-            updates.write_to_database(provider.tx_ref())?;
+            updates.write_to_database(&TxRefWrapper::from(provider.tx_ref()))?;
 
             let total_hashed_entries = (provider.count_entries::<tables::HashedAccounts>()? +
                 provider.count_entries::<tables::HashedStorages>()?)
@@ -325,7 +326,7 @@ impl<DB: Database> Stage<DB> for MerkleStage {
             validate_state_root(block_root, target.seal_slow(), input.unwind_to)?;
 
             // Validation passed, apply unwind changes to the database.
-            updates.write_to_database(provider.tx_ref())?;
+            updates.write_to_database(&TxRefWrapper::from(provider.tx_ref()))?;
 
             // TODO(alexey): update entities checkpoint
         } else {
