+++ crates/storage/provider/src/providers/database/provider.rs
@@ -46,6 +46,7 @@ use reth_trie::{
     updates::TrieUpdates,
     HashedPostState, Nibbles, StateRoot,
 };
+use reth_trie_db::{trie::StateRootDb, TxRefWrapper};
 use revm::primitives::{BlockEnv, CfgEnvWithHandlerCfg};
 use std::{
     cmp::Ordering,
@@ -2441,7 +2442,7 @@ impl<TX: DbTxMut + DbTx> HashingWriter for DatabaseProvider<TX> {
                     block_hash: end_block_hash,
                 })))
             }
-            trie_updates.write_to_database(&self.tx)?;
+            trie_updates.write_to_database(&TxRefWrapper::from(&self.tx))?;
         }
         durations_recorder.record_relative(metrics::Action::InsertMerkleTree);
 
@@ -2637,7 +2638,7 @@ impl<TX: DbTxMut + DbTx> BlockExecutionWriter for DatabaseProvider<TX> {
                     block_hash: parent_hash,
                 })))
             }
-            trie_updates.write_to_database(&self.tx)?;
+            trie_updates.write_to_database(&TxRefWrapper::from(&self.tx))?;
         }
 
         // get blocks
@@ -2840,7 +2841,7 @@ impl<TX: DbTxMut + DbTx> BlockWriter for DatabaseProvider<TX> {
         // insert hashes and intermediate merkle nodes
         {
             HashedStateChanges(hashed_state).write_to_db(&self.tx)?;
-            trie_updates.write_to_database(&self.tx)?;
+            trie_updates.write_to_database(&TxRefWrapper::from(&self.tx))?;
         }
         durations_recorder.record_relative(metrics::Action::InsertHashes);
