+++ crates/trie/db/src/cursors/hashed.rs
@@ -1,43 +1,47 @@
-use super::{HashedCursor, HashedCursorFactory, HashedStorageCursor};
+use crate::{DbCursorWrapper, TxRefWrapper};
 use reth_db::tables;
 use reth_db_api::{
     cursor::{DbCursorRO, DbDupCursorRO},
     transaction::DbTx,
+    DatabaseError,
 };
 use reth_primitives::{Account, B256, U256};
+use reth_trie::hashed_cursor::{HashedCursor, HashedCursorFactory, HashedStorageCursor};
 
-impl<'a, TX: DbTx> HashedCursorFactory for &'a TX {
-    type AccountCursor = <TX as DbTx>::Cursor<tables::HashedAccounts>;
+impl<'a, TX: DbTx> HashedCursorFactory for TxRefWrapper<'a, TX> {
+    type Err = DatabaseError;
+    type AccountCursor = DbCursorWrapper<<TX as DbTx>::Cursor<tables::HashedAccounts>>;
     type StorageCursor =
         DatabaseHashedStorageCursor<<TX as DbTx>::DupCursor<tables::HashedStorages>>;
 
-    fn hashed_account_cursor(&self) -> Result<Self::AccountCursor, reth_db::DatabaseError> {
-        self.cursor_read::<tables::HashedAccounts>()
+    fn hashed_account_cursor(&self) -> Result<Self::AccountCursor, Self::Err> {
+        Ok(DbCursorWrapper(self.0.cursor_read::<tables::HashedAccounts>()?))
     }
 
     fn hashed_storage_cursor(
         &self,
         hashed_address: B256,
-    ) -> Result<Self::StorageCursor, reth_db::DatabaseError> {
+    ) -> Result<Self::StorageCursor, Self::Err> {
         Ok(DatabaseHashedStorageCursor::new(
-            self.cursor_dup_read::<tables::HashedStorages>()?,
+            self.0.cursor_dup_read::<tables::HashedStorages>()?,
             hashed_address,
         ))
     }
 }
 
-impl<C> HashedCursor for C
+impl<C> HashedCursor for DbCursorWrapper<C>
 where
     C: DbCursorRO<tables::HashedAccounts>,
 {
+    type Err = DatabaseError;
     type Value = Account;
 
-    fn seek(&mut self, key: B256) -> Result<Option<(B256, Self::Value)>, reth_db::DatabaseError> {
-        self.seek(key)
+    fn seek(&mut self, key: B256) -> Result<Option<(B256, Self::Value)>, Self::Err> {
+        self.0.seek(key)
     }
 
-    fn next(&mut self) -> Result<Option<(B256, Self::Value)>, reth_db::DatabaseError> {
-        self.next()
+    fn next(&mut self) -> Result<Option<(B256, Self::Value)>, Self::Err> {
+        self.0.next()
     }
 }
 
@@ -63,16 +67,14 @@ impl<C> HashedCursor for DatabaseHashedStorageCursor<C>
 where
     C: DbCursorRO<tables::HashedStorages> + DbDupCursorRO<tables::HashedStorages>,
 {
+    type Err = DatabaseError;
     type Value = U256;
 
-    fn seek(
-        &mut self,
-        subkey: B256,
-    ) -> Result<Option<(B256, Self::Value)>, reth_db::DatabaseError> {
+    fn seek(&mut self, subkey: B256) -> Result<Option<(B256, Self::Value)>, Self::Err> {
         Ok(self.cursor.seek_by_key_subkey(self.hashed_address, subkey)?.map(|e| (e.key, e.value)))
     }
 
-    fn next(&mut self) -> Result<Option<(B256, Self::Value)>, reth_db::DatabaseError> {
+    fn next(&mut self) -> Result<Option<(B256, Self::Value)>, Self::Err> {
         Ok(self.cursor.next_dup_val()?.map(|e| (e.key, e.value)))
     }
 }
@@ -81,7 +83,7 @@ impl<C> HashedStorageCursor for DatabaseHashedStorageCursor<C>
 where
     C: DbCursorRO<tables::HashedStorages> + DbDupCursorRO<tables::HashedStorages>,
 {
-    fn is_storage_empty(&mut self) -> Result<bool, reth_db::DatabaseError> {
+    fn is_storage_empty(&mut self) -> Result<bool, Self::Err> {
         Ok(self.cursor.seek_exact(self.hashed_address)?.is_none())
     }
 }
