+++ crates/trie/parallel/src/parallel_root.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "metrics")]
+use crate::metrics::ParallelStateRootMetrics;
 use crate::{stats::ParallelTrieTracker, storage_root_targets::StorageRootTargets};
 use alloy_rlp::{BufMut, Encodable};
 use rayon::prelude::*;
@@ -13,12 +15,10 @@ use reth_trie::{
     walker::TrieWalker,
     HashBuilder, HashedPostState, Nibbles, StorageRoot, TrieAccount,
 };
+use reth_trie_db::TxRefWrapper;
 use std::collections::HashMap;
 use thiserror::Error;
 use tracing::*;
-
-#[cfg(feature = "metrics")]
-use crate::metrics::ParallelStateRootMetrics;
 
 /// Parallel incremental state root calculator.
 ///
@@ -91,9 +91,10 @@ where
             .into_par_iter()
             .map(|(hashed_address, prefix_set)| {
                 let provider_ro = self.view.provider_ro()?;
+                let factory = TxRefWrapper::from(provider_ro.tx_ref());
                 let storage_root_result = StorageRoot::new_hashed(
-                    provider_ro.tx_ref(),
-                    HashedPostStateCursorFactory::new(provider_ro.tx_ref(), &hashed_state_sorted),
+                    factory,
+                    HashedPostStateCursorFactory::new(factory, &hashed_state_sorted),
                     hashed_address,
                     #[cfg(feature = "metrics")]
                     self.metrics.storage_trie.clone(),
@@ -108,9 +109,10 @@ where
         let mut trie_updates = TrieUpdates::default();
 
         let provider_ro = self.view.provider_ro()?;
+        let factory = TxRefWrapper::from(provider_ro.tx_ref());
         let hashed_cursor_factory =
-            HashedPostStateCursorFactory::new(provider_ro.tx_ref(), &hashed_state_sorted);
-        let trie_cursor_factory = provider_ro.tx_ref();
+            HashedPostStateCursorFactory::new(factory, &hashed_state_sorted);
+        let trie_cursor_factory = factory;
 
         let walker = TrieWalker::new(
             trie_cursor_factory.account_trie_cursor().map_err(ProviderError::Database)?,
