+++ crates/trie/trie/src/trie_cursor/noop.rs
@@ -1,6 +1,5 @@
 use super::{TrieCursor, TrieCursorFactory};
 use crate::{BranchNodeCompact, Nibbles};
-use reth_db::DatabaseError;
 use reth_primitives::B256;
 
 /// Noop trie cursor factory.
@@ -9,19 +8,17 @@ use reth_primitives::B256;
 pub struct NoopTrieCursorFactory;
 
 impl TrieCursorFactory for NoopTrieCursorFactory {
+    type Err = ();
     type AccountTrieCursor = NoopAccountTrieCursor;
     type StorageTrieCursor = NoopStorageTrieCursor;
 
     /// Generates a Noop account trie cursor.
-    fn account_trie_cursor(&self) -> Result<Self::AccountTrieCursor, DatabaseError> {
+    fn account_trie_cursor(&self) -> Result<Self::AccountTrieCursor, ()> {
         Ok(NoopAccountTrieCursor::default())
     }
 
     /// Generates a Noop storage trie cursor.
-    fn storage_trie_cursor(
-        &self,
-        _hashed_address: B256,
-    ) -> Result<Self::StorageTrieCursor, DatabaseError> {
+    fn storage_trie_cursor(&self, _hashed_address: B256) -> Result<Self::StorageTrieCursor, ()> {
         Ok(NoopStorageTrieCursor::default())
     }
 }
@@ -32,24 +29,20 @@ impl TrieCursorFactory for NoopTrieCursorFactory {
 pub struct NoopAccountTrieCursor;
 
 impl TrieCursor for NoopAccountTrieCursor {
+    type Err = ();
+
     /// Seeks an exact match within the account trie.
-    fn seek_exact(
-        &mut self,
-        _key: Nibbles,
-    ) -> Result<Option<(Nibbles, BranchNodeCompact)>, DatabaseError> {
+    fn seek_exact(&mut self, _key: Nibbles) -> Result<Option<(Nibbles, BranchNodeCompact)>, ()> {
         Ok(None)
     }
 
     /// Seeks within the account trie.
-    fn seek(
-        &mut self,
-        _key: Nibbles,
-    ) -> Result<Option<(Nibbles, BranchNodeCompact)>, DatabaseError> {
+    fn seek(&mut self, _key: Nibbles) -> Result<Option<(Nibbles, BranchNodeCompact)>, ()> {
         Ok(None)
     }
 
     /// Retrieves the current cursor position within the account trie.
-    fn current(&mut self) -> Result<Option<Nibbles>, DatabaseError> {
+    fn current(&mut self) -> Result<Option<Nibbles>, ()> {
         Ok(None)
     }
 }
@@ -60,24 +53,20 @@ impl TrieCursor for NoopAccountTrieCursor {
 pub struct NoopStorageTrieCursor;
 
 impl TrieCursor for NoopStorageTrieCursor {
+    type Err = ();
+
     /// Seeks an exact match in storage tries.
-    fn seek_exact(
-        &mut self,
-        _key: Nibbles,
-    ) -> Result<Option<(Nibbles, BranchNodeCompact)>, DatabaseError> {
+    fn seek_exact(&mut self, _key: Nibbles) -> Result<Option<(Nibbles, BranchNodeCompact)>, ()> {
         Ok(None)
     }
 
     /// Seeks a key in storage tries.
-    fn seek(
-        &mut self,
-        _key: Nibbles,
-    ) -> Result<Option<(Nibbles, BranchNodeCompact)>, DatabaseError> {
+    fn seek(&mut self, _key: Nibbles) -> Result<Option<(Nibbles, BranchNodeCompact)>, ()> {
         Ok(None)
     }
 
     /// Retrieves the current cursor position within storage tries.
-    fn current(&mut self) -> Result<Option<Nibbles>, DatabaseError> {
+    fn current(&mut self) -> Result<Option<Nibbles>, ()> {
         Ok(None)
     }
 }
